Up to Exercise 9.25
Error messages are displayed if some required values are missing or formatted incorrectly. This is achieved through utils.ts, which throws errors. The thrown errors are caught by try-catch in the patient router, which is sent to the PatientDetail frontend through another try-catch. The error state is set to the message that was propagated from utils.ts.